requirements.txt
Django==5.2
djangorestframework==3.15.1
pytz
sqlparse
asgiref
requests
icalendar
psycopg2-binary==2.9.9